Allowed use of code flow without PKCE and removed client secret from code
